[front] refactor(skills): break circular dep skill_resource -> agent -> global_agents
Extract updateAgentRequirements to agent_requirements.ts and replace
getAgentConfiguration with fetchMCPServerActionConfigurations + new
listByAgentModelId so skill_resource.ts no longer imports from
configuration/agent.ts (which transitively pulls in global_agents.ts).
