+++ front/lib/resources/skill/skill_resource.ts
@@ -1,7 +1,5 @@
-import {
-  getAgentConfiguration,
-  updateAgentRequirements,
-} from "@app/lib/api/assistant/configuration/agent";
+import { fetchMCPServerActionConfigurations } from "@app/lib/actions/configuration/mcp";
+import { updateAgentRequirements } from "@app/lib/api/assistant/configuration/agent_requirements";
 import { getAgentConfigurationRequirementsFromCapabilities } from "@app/lib/api/assistant/permissions";
 import { hasSharedMembership } from "@app/lib/api/user";
 import type { Authenticator } from "@app/lib/auth";
@@ -1229,6 +1227,24 @@ export class SkillResource extends BaseResource<SkillConfigurationModel> {
     return this.editorGroup.canWrite(auth);
   }
 
+  private async listOtherAgentSkills(
+    auth: Authenticator,
+    agent: AgentConfigurationModel
+  ): Promise<SkillResource[]> {
+    const workspace = auth.getNonNullableWorkspace();
+    const agentSkills = await AgentSkillModel.findAll({
+      where: { agentConfigurationId: agent.id, workspaceId: workspace.id },
+    });
+    const skills = await SkillResource.fetchBySkillReferences(
+      auth,
+      agentSkills.map((s) => ({
+        customSkillId: s.customSkillId,
+        globalSkillId: s.globalSkillId,
+      }))
+    );
+    return skills.filter((skill) => skill.sId !== this.sId);
+  }
+
   private async listActiveAgents(
     auth: Authenticator
   ): Promise<AgentConfigurationModel[]> {
@@ -1302,23 +1318,17 @@ export class SkillResource extends BaseResource<SkillConfigurationModel> {
         // removed from the agent. In order to achieve this, we check if the agent has
         // any other capabilities that require the removed spaces.
         if (spaceIdsRemovedFromThisSkill.length > 0) {
-          const agentConfig = await getAgentConfiguration(auth, {
-            agentId: agent.sId,
+          const actionsMap = await fetchMCPServerActionConfigurations(auth, {
+            configurationIds: [agent.id],
             variant: "full",
           });
-          assert(agentConfig, "Agent configuration not found");
+          const actions = actionsMap.get(agent.id) ?? [];
 
-          const agentSkills = await SkillResource.listByAgentConfiguration(
-            auth,
-            agentConfig
-          );
-          const otherAgentSkills = agentSkills.filter(
-            (skill) => skill.sId !== this.sId
-          );
+          const otherAgentSkills = await this.listOtherAgentSkills(auth, agent);
 
           const agentOtherCapabilitiesRequirements =
             await getAgentConfigurationRequirementsFromCapabilities(auth, {
-              actions: agentConfig.actions,
+              actions,
               skills: otherAgentSkills,
             });
